:sparkles: (Board And Orb): Added orb connection evaluation
Orbs that are dropped now evaluate themselfs and others around them to see if there are 3 or more connected orbs, BUG orbs in any direction and order will pop if you have 3 touching this should be changed to vertical only if we are going for a magical drop style game. Orbs currently are not destroied and basic code to remove them from the que's they are in as well as to remove them from the game should be added. as a side thing I was using linked lists backwards so a lot of Firsts and Lasts have been fliped